refactor-#35: move activity repository to activity
