+++ src/catalog/activity/queries/handlers/get-activities-by-query.handler.ts
@@ -4,7 +4,7 @@ import { In } from 'typeorm';
 
 import { EsService } from '../../../../es/es.service';
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { ActivityRepository } from '../../../../db/repositories';
+import { ActivityRepository } from '../../repositories';
 import { GetActivitiesByQueryQuery } from '../impl';
 import { ActivityEntity } from '../../../../db/entities/activity.entity';
